+++ src/ui/components/index.ts
@@ -3,4 +3,4 @@ import "./label";
 import "./option";
 import "./radio-group";
 import "./switch";
-import "./textfield";
+import "./text-field";
